chore: harden release and quality gates
- Add Prettier, ESLint, Node 22, and unified make/pnpm checks
- Expand CI with formatting, linting, coverage, Clippy, audit, and release checks
- Rework macOS release tooling for community and notarized release modes
- Verify updater signatures and tighten build/notarization validation
- Convert bundled JetBrains Mono fonts to WOFF2 and document licensing

+++ src-tauri/src/aws/billing.rs
@@ -169,7 +169,7 @@ fn ymd_utc(secs: u64) -> (i32, u32, u32) {
     let days = (secs / 86400) as i64;
     let z = days + 719468;
     let era = if z >= 0 { z } else { z - 146096 } / 146097;
-    let doe = (z - era * 146097) as i64;
+    let doe = z - era * 146097;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
